refactor(import referentiel): extract csv parsing
